Heatmap: color ramp research
